final commit

+++ database/migrations/2019_10_15_091050_create_orders_table.php
@@ -18,7 +18,7 @@ class CreateOrdersTable extends Migration
             $table->unsignedBigInteger('customer_id');
             $table->string('title');
             $table->text('description');
-            $table->bigInteger('cost');
+            $table->double('cost', 8, 2);
             $table->timestamps();
             $table->softDeletes();
